refactor: move PartialRowsData testing to test_row_data.py
